fix(building-webpack): improve setup docu to minify user errors
Added notes to avoid problems I ran into as a Polymer CLI user using webpack for the first time.

Also moved the npm scripts to the last step. When I hit this part of the instructions, I immediately tried running the scripts, which failed because the necessary HTML and JS files weren't set up yet. Moving this step to the end prevents this user error.

Changed one header to make it more descriptive.